requirements.txt
databases~=0.8.0
SQLAlchemy~=1.4.50
pytest~=7.4.3
fastapi~=0.104.1
pydantic~=2.5.2
uvicorn~=0.24.0.post1
starlette~=0.27.0
python-dotenv~=1.0.0